+++ src/main/java/com/fatec/backend/repository/LivroRepository.java
@@ -4,6 +4,8 @@ import com.fatec.backend.model.Livro;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface LivroRepository extends JpaRepository<Livro, Integer> {
+public interface LivroRepository extends JpaRepository<Livro, Long> {
 }
